usb: i.MX: warn if vbus regulator isn't available
Instead of just ignoring errors related to getting the vbus regulator
yield at least a warning message. This would have saved me some
debugging time when trying to understand why USB doesn't work without
the right regulator driver enabled.

Note that machines that don't define a regulator in their device tree
don't issue this warning as regulator_get() returns the dummy regulator
in this case.

The alternative to error out wasn't accepted because this probably
creates regressions for regulators that are default-on but without a
driver in barebox.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Link: https://lore.barebox.org/20210305090001.3619780-1-u.kleine-koenig@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>