ARM64: qemu-virt64: remove dead CPU identification code
cpu_is_arm(cpu) masks the read_cpuid_id() return value and compares with
a CPU-specific value. The used masks for the Cortex A53/57 zeroes bits
that are set in the value, so the comparison will never return true.

In fact the compiler already optimizes it away completely, because
read_cpuid_id() is only implemented for ARM32 anyway.

Drop the dead code.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>