ARM64: board-dt-2nd: adopt kernel image header
The ARM64 generic DT image is meant to be interchangeable with a
kernel image to aid in debugging and testing. To achieve this, it
reuses an externally passed device tree in x0, just like Linux does.

The ARM barebox image format used however imposes some limitations:

  - Commands verifying the header before boot, like U-Boot's booti
    won't boot the barebox image

  - The barebox image is not fully relocatable. It has the implicit
    assumption that the barebox stack can grow into the memory space
    preceding the barebox image while the /memory node is parsed
    from the FDT.

Adopting the Linux ARM64 header solves both issues. booti won't be
able to tell us apart and we can specify an image load offset to
reserve a stack space that won't interfere with anything else.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>