blspec: Check if already mounted before mounting
cdev_mount_default() either mounts a cdev and returns the path it's
mounted to or directly returns the path when it already is mounted.
This will change in the next patch: When the cdev is already mounted
cdev_mount_default() will return -EBUSY instead. To let blspec
work when a cdev is already mounted check for the actual mount path
before trying to mount it.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>