scripts: imx: add support for additional hab Blocks
This allows to specifiy additional signed blocks
in the format `offset+size@address` within the imximg.

It is needed by the uuu tool, which loads the image different to
the imx-usb-loader. It loads the DCD always to 0x910000 in the OCRAM.
So this area have to be signed as well.

In my case this needs `hab_blocks 0x42c+0x1f0@0x910000`.

It supports to remove quotes to support Kconfig variable here.

Signed-off-by: Denis Osterland-Heim <denis.osterland@diehl.com>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>