ARM: pbl: uncompress: add comment explaining order of operations
When the PBL runs from flash, it relocates to RAM prior to extracting
barebox proper. It does the sneaky thing of _not_ relocating the piggy
data though, so read from flash and uncompression may latter happen
at the same time.

For this to work, it's critical that the variables pointing at the piggy
data are evaluated before relocation. ARM does so and carefully uses
them to derive the address of the PBL size later on instead of
evaluating image_data_end again.

When I ported the same code for RISC-V use, this got lost. Add a comment
explaining the criticality of maintaining this order of operations.

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>
Link: https://lore.pengutronix.de/20210403070346.3643335-1-ahmad@a3f.at
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>