sandbox: env: set nv.autoboot=abort to drop to shell
barebox on sandbox defaults to network boot, which fails either due to
missing TAP support (like not having CAP_NET_ADMIN) or because the files
couldn't be fetched.

Arguably, the most user-friendly thing is to drop to shell. Do so.
Users can still override nv.autoboot if they want to boot differently.

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>
Acked-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Link: https://lore.pengutronix.de/20210404111831.592821-1-ahmad@a3f.at
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>