common: add option for shipping KCONFIG_CONFIG as /env/data/config
Add a new option that enables access to the barebox configuration
file through /env/data/config. This can be useful in environments
where a fatter barebox can be tolerated (like on an EFI system
partition). It could also be used to conditionally skip tests
on disabled functionality in the future.

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>
Tested-by: Roland Hieber <rhi@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>