watchdog: add support for Intel TCO watchdog timer
Variants of the iTCO are integrated into many Intel southbridges.
They are most often accessed via PCI. Add a driver for the variant
found in the q35 QEMU machine.

It should be straight forward to extend the itco_chipset_info array
to support more variants in future as the need arises. To test, use:

  qemu-system-x86_64 -M q35 -global ICH9-LPC.noreboot=false

The last option corresponds to a pin strap option, which can't be
influenced from within the VM.

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>
Link: https://lore.barebox.org/20210416062436.332665-5-ahmad@a3f.at
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>