usb: storage: retry for up to 10s on lengthy HDD spin up
Some USB disks take notoriously long to spin up. They are seen by a bus
scan, but they report ready only after a few seconds have passed. This
is not a problem if vbus is enabled early on, so devices have had a
chance to spin up. If vbus is first enabled as part of the usb scan,
not enough time might have passed for the USB disk to be usable.

This issue was observed on an i.MX6QP with following topology:

  usb: USB: scanning bus for devices...
  usb: 5 USB Device(s) found
    1 ID 0000:0000
    |  u-boot EHCI Host Controller
    |
    +-2 ID 0424:2517
    |
    +-5 ID 1058:2621
    |    Western Digital Elements 2621
    ...

Unplugging and replugging the USB disk and doing a second usb scan
made the unit ready test succeed. Increasing the retry count
during initialization has negative consequences for other cases,
like when a device is unplugged while being probed (which already
takes way too long).

Instead, just for the case of a detected USB mass storage device that
couldn't get ready initially: retry for 10s at initialization time
before giving up.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20210416172646.26834-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>