ARM: i.MX: demote reset reason print to debug level
The i.MX reset reason is not really helpful on its own, as it needs
to be augmented with other information like the watchdog state to get
the real system reset reason. As it stands this log is more confusing
than answering any questions a user might have.

Demote the log to debug level to stop this confusion.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>
Reviewed-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20210422110849.1556472-1-l.stach@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>