usb: dwc2: Fix wait for mode timeout
The timeout waiting for host mode should be 110ms instead of 110us.
The timeout is the same as in Linux now and the same it was before
fdd30cc251. It takes 49ms to go to host mode on the Raspberry Pi 3b,
so host mode stopped working with 26459ab780 which removed an additional
100ms delay before dwc2_wait_for_mode() was actually called.

Fixes: fdd30cc251 ("usb: dwc2: Rework dwc2_wait_for_mode to use wait_on_timeout")
Fixes: 26459ab780 ("usb: dwc2: Rework wait for host mode during core reset")
Reported-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>