usb: dwc2: more precise and helpful message for dr_mode config issues
With the previous message I misunderstood that the 'configuration
mismatch' was caused by any entity forcing the dr_mode to
'host'/'device'.
The actual intention however is to tell the user that the selected
'dr_mode' does not match the capabilities provided by the controller or
the selected driver parts (USB_DWC2_HOST/USB_DWC2_GADGET).

The updated warning message attempts to reflect this more explicitly.

Also rename 'device' to 'peripheral' as it is named both in dtb and
in macros that way.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>
Acked-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20210517124941.31301-1-ejo@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>