common: memory: allocate all memory devices at once
Follow-up commit will fuse overlapping RAM banks. As all memory is
supposed to be registered during mem_initcall or before, we can postpone
device creation to mmu_initcall, so we can directly allocate devices
spanning the correct region.

The mem driver and the devinfo command are the only consumers of these
devices, so it's ok to register the devices at mmu_initcall.

While at it, drop the struct memory_bank::dev member.
It's unused anywhere.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20210531071239.30653-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>