memory: fuse overlapping memory banks
Some ARM subarchitectures read back RAM size from the SDRAM
controller and use the info to register memory banks.

If an overlapping memory region is already registered, e.g. via device
tree /memory, this second registration will fail.

This is especially annoying as it can regress after a device tree sync:

 - Kind soul updates upstream device tree to describe minimal available
   RAM across hardware variants
 - barebox PBL has enough info about the board to set up larger RAM size
   and relocates barebox to the end of the RAM
 - barebox proper starts with new device tree and is upset to
   find itself outside of registered memory

Account for this by growing the existing bank if a bank to be added
happens to overlap it. As a special case, if the existing bank
completely contains the new memory bank, the function is a no-op.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20210531071239.30653-3-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>