drivers: base: Check all compatible strings for modalias match
When attempting a modalias match in device_match_of_modalias(), only the
first string in the compatible property, which is a list of strings, was
used.

A modalias (which is a bit of a misnomer in Barebox) match is used when
a driver does not have an of_compatible match table, e.g. the at24
driver.  The compatible string after the comma is matched against the
driver's id table.

Extend modalias match to try all strings in from the OF node's
compatible property.  This will cause a compatible like
"rohm,br24g04-3", "atmel,24c04" to match against the "24c04" ID in the
at24 driver.  Or "isil,isl12057", "dallas,ds1337" will match the ds1307
driver's table, which doesn't know about isl12057 in Barebox.

Signed-off-by: Trent Piepho <tpiepho@gmail.com>
Link: https://lore.barebox.org/20210531145414.1210207-1-tpiepho@gmail.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>