fs: nfs: Start earlier to resend requests
On a customer site we're experience a bit over 1% UDP packet loss. When
wiresharking an NFS transfer of a kernel image (with the goal to boot
via NFS) I saw 64 of 2555 RPC calls staying unanswered. With the current
timeout setting each of them introduces a delay of 2 seconds and the
whole transfer takes 137s. With the timeout reduced to 0.1s the transfer
time is not optimal (going down to approx 15 seconds) but at least it
becomes bearable.

To still cope with a slow network (in contrast to an unreliable as
pictured above) increase NFS_MAX_RESEND to keep NFS_TIMEOUT *
NFS_MAX_RESEND (which defines the overall timeout before aborting the
transfer) constant.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Acked-by: Gavin Schenk <g.schenk@eckelmann.de>
Link: https://lore.barebox.org/20210611071351.1445370-1-u.kleine-koenig@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>